requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow>=10.3.0
numpy>=1.26.0
scikit-learn>=1.4.0
requests==2.31.0
